# artifacts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Taobei App API",
    "version": "1.1.0",
    "description": "淘贝应用后端接口规范，覆盖登录、注册与验证码相关场景（错误字段统一为 error）"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "本地开发服务器" }
  ],
  "paths": {
    "/auth/request-code": {
      "post": {
        "summary": "请求验证码（登录/注册通用）",
        "description": "校验手机号格式，生成6位验证码，打印到控制台并在数据库中记录（有效期60秒）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PhoneNumberRequest" },
              "examples": {
                "valid": { "value": { "phone": "13800138000" } },
                "invalid": { "value": { "phone": "123" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "验证码生成成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RequestCodeResponse" },
                "examples": {
                  "success": { "value": { "message": "验证码已发送", "seconds": 60 } }
                }
              }
            }
          },
          "400": {
            "description": "手机号格式无效",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "invalidPhone": { "value": { "error": "请输入正确的手机号码" } }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "登录",
        "description": "校验手机号是否已注册、验证码是否正确且未过期；成功则返回提示消息，前端可在提示后跳转首页",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" },
              "examples": {
                "valid": { "value": { "phone": "13800138000", "code": "123456" } },
                "unregistered": { "value": { "phone": "13100000000", "code": "123456" } },
                "wrongCode": { "value": { "phone": "13800138000", "code": "000000" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" },
                "examples": {
                  "success": { "value": { "message": "登录成功" } }
                }
              }
            }
          },
          "400": {
            "description": "错误场景：未注册/验证码错误/验证码已过期/手机号无效",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "invalidPhone": { "value": { "error": "请输入正确的手机号码" } },
                  "unregistered": { "value": { "error": "该手机号未注册，请先完成注册" } },
                  "wrongCode": { "value": { "error": "验证码错误" } },
                  "expired": { "value": { "error": "验证码已过期" } }
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "注册",
        "description": "校验手机号与验证码、协议勾选；已注册则直接提示并视为登录成功，未注册则创建用户并提示注册成功",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" },
              "examples": {
                "valid": { "value": { "phone": "13800138000", "code": "123456", "agree": true } },
                "notAgreed": { "value": { "phone": "13800138000", "code": "123456", "agree": false } },
                "wrongCode": { "value": { "phone": "13800138000", "code": "000000", "agree": true } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功或已注册直接登录",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" },
                "examples": {
                  "created": { "value": { "message": "注册成功" } },
                  "already": { "value": { "message": "该手机号已注册，将直接为您登录" } }
                }
              }
            }
          },
          "400": {
            "description": "错误场景：手机号无效/未同意协议/验证码错误/验证码已过期",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "invalidPhone": { "value": { "error": "请输入正确的手机号码" } },
                  "notAgreed": { "value": { "error": "请先同意《淘贝用户协议》" } },
                  "wrongCode": { "value": { "error": "验证码错误" } },
                  "expired": { "value": { "error": "验证码已过期" } }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PhoneNumberRequest": {
        "type": "object",
        "required": ["phone"],
        "properties": {
          "phone": { "type": "string", "pattern": "^1\\d{10}$", "example": "13800138000" }
        }
      },
      "RequestCodeResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "验证码已发送" },
          "seconds": { "type": "integer", "example": 60 }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["phone", "code"],
        "properties": {
          "phone": { "type": "string", "pattern": "^1\\d{10}$", "example": "13800138000" },
          "code": { "type": "string", "minLength": 6, "maxLength": 6, "example": "123456" }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["phone", "code", "agree"],
        "properties": {
          "phone": { "type": "string", "pattern": "^1\\d{10}$", "example": "13800138000" },
          "code": { "type": "string", "minLength": 6, "maxLength": 6, "example": "123456" },
          "agree": { "type": "boolean", "example": true }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  }
}